move authentication part into javascript folder
