+++ src/javascripts/app.js
@@ -55,13 +55,13 @@ require('./common/serverApi.service');
 require('./common/uid.service.js');
 require('./common/previousState.service.js');
 require('./common/unsubscribeOnPageChange.service.js');
-require('./common/authentication/authentication.module');
-require('./common/authentication/authentication.service');
-require('./common/authentication/loginModal.controller');
-require('./common/authentication/authorizationApi.service');
-require('./common/authentication/login.controller');
-require('./common/authentication/formAutoFillFix.directive');
-require('./common/authentication/userMenu.directive');
+require('./authentication/authentication.module');
+require('./authentication/authentication.service');
+require('./authentication/loginModal.controller');
+require('./authentication/authorizationApi.service');
+require('./authentication/login.controller');
+require('./authentication/formAutoFillFix.directive');
+require('./authentication/userMenu.directive');
 require('./common/headline/headline.directive');
 
 import baseRouting from './baseRouting';
