+++ src/javascripts/authentication/authEvents.const.js
