+++ src/javascripts/authentication/login.controller.js
