+++ src/javascripts/authentication/loginModal.controller.js
