+++ src/javascripts/authentication/session.service.js
