+++ src/javascripts/profile/browse.ctrl.js
@@ -3,7 +3,7 @@ import cogOptionsCollection from '../collection/cogOptionsCollection/cogOptionsC
 import documentsInline from '../common/documentsInline/documentsInline.directive';
 import profileToolbar from '../common/documentsInline/profileToolbar.directive';
 import profileApi from '../common/profileApi.service';
-import authorizationApi from '../common/authentication/authorizationApi.service';
+import authorizationApi from '../authentication/authorizationApi.service';
 
 const ctrlName = 'ProfileBrowseCtrl';
