+++ src/javascripts/user/full.ctrl.js
@@ -3,7 +3,7 @@ import addAttribute from '../storage/addAttribute/addAttribute.directive';
 import userApi from '../common/userApi.service';
 import uiNotification from 'angular-ui-notification';
 import prevousState from '../common/previousState.service.js';
-import authorizationApi from '../common/authentication/authorizationApi.service';
+import authorizationApi from '../authentication/authorizationApi.service';
 
 var ctrlName = 'UserFullCtrl';
